chore: rename migrations.ts to migration.ts (migration engine)
