+++ packages/keyring-sdk/src/index.ts
@@ -1,5 +1,5 @@
 export * from './keyring-wrapper';
 export * from './keyring-account-registry';
-export * from './migrations';
+export * from './migration';
 export * from './mnemonic';
 export * from './eth';
